requirements.txt
customtkinter==5.2.2
googletrans==4.0.0rc1
speechrecognition==3.10.3
pyttsx3==2.90
pyaudio==0.2.14
setuptools==69.2.0
pillow==10.3.0
opencv-python==4.9.0.80
pyvirtualcam==0.11.1
pygrabber==0.2
pywin32